requirements.txt
flask==3.1.0
rembg==2.0.65
Pillow==10.2.0
python-dotenv==1.0.1
opencv-python==4.9.0.80
numpy==1.26.4
werkzeug>=2.0.1 